+++ mch2022/components/mch2022_rp2040/mch2022_rp2040.h
@@ -53,11 +53,11 @@ class Mch2022_rp2040Component : public Component, public i2c::I2CDevice {
       RP2040_INPUT_JOYSTICK_RIGHT,
       SUB_BINARY_SENSOR_INPUT_COUNT
     };
-  void set_sub_binary_sensor(size_t input, binary_sensor::BinarySensor *sens);
+  void set_sub_binary_sensor(uint8_t input, binary_sensor::BinarySensor *sens);
 
  protected:
-  void update_sub_binary_sensor_(size_t input, bool value);
-  std::map<size_t, *binary_sensor::BinarySensor> sub_binary_sensors_;
+  void update_sub_binary_sensor_(uint8_t input, bool value);
+  std::map<uint8_t, *binary_sensor::BinarySensor> sub_binary_sensors_;
   // binary_sensor::BinarySensor *sub_binary_sensors_[(size_t) SubBinarySensorInput::SUB_BINARY_SENSOR_INPUT_COUNT]{nullptr};
 #endif
   public:
